Fix checks in timing

+++ jenkins-scripts/docker/lib/_generic_linux_compilation_build.sh.bash
@@ -38,8 +38,10 @@ echo '# END SECTION'
 
 echo '# BEGIN SECTION: cppcheck'
 cd $WORKSPACE/${SOFTWARE_DIR}
-mkdir -p tools
-cp $WORKSPACE/scripts/jenkins-scripts/tools/cpplint_to_cppcheckxml.py tools/
+if [ ! -f tools/cpplint_to_cppcheckxml.py ]; then
+  mkdir -p tools
+  cp $WORKSPACE/scripts/jenkins-scripts/tools/cpplint_to_cppcheckxml.py tools/
+fi
 init_stopwatch CPPCHECK
 sh tools/code_check.sh -xmldir $WORKSPACE/build/cppcheck_results || true
 stop_stopwatch CPPCHECK
